+++ VendingMachine/src/View/UserIOInterface.java
@@ -16,8 +16,8 @@ public interface UserIOInterface {
     public void print(String msgPrompt);
     public String readString(String msgPrompt);
     
-    public BigDecimal readDouble(String msgPrompt);
-    public BigDecimal readDouble(String msgPrompt, BigDecimal min);
+    public BigDecimal readBigDecimal(String msgPrompt);
+    public BigDecimal readBigDecimal(String msgPrompt, BigDecimal min);
     
     public int readInt(String msgPrompt);
     public int readInt(String msgPrompt, int min, int max);
